requirements.txt
numpy==1.24.2
torch==2.1.0
torchvision==0.16.0
pyyaml==6.0.1
pillow==10.0.1
matplotlib==3.8.1
CairoSVG==2.7.1
Flask==3.0.0
scipy==1.11.4
opencv-python==4.8.1.78
gdown==4.7.1

